requirements.txt
fastapi>=0.100.0
pydantic>=2.0.0
streamlit>=1.0.0
requests>=2.25.0
pandas>=1.3.0
langgraph>=0.0.40
python-dotenv>=0.19.0
langchain>=0.2.0
langchain-groq>=0.1.0
langchain-community>=0.2.0
